break panel into new `GuiPanelModels` + organize
