+++ src/components/panel/GuiPanel.tsx
@@ -1,46 +1,12 @@
-import {
-  selectModel,
-  selectModelCount,
-  selectModelIndex,
-  selectObjectKey,
-  selectObjectSelectionType,
-  setObjectType,
-  useAppDispatch,
-  useAppSelector,
-  navToPrevModel,
-  navToNextModel
-} from '@/store';
-import {
-  Button,
-  IconButton,
-  Paper,
-  ToggleButton,
-  ToggleButtonGroup,
-  Tooltip,
-  Typography,
-  styled
-} from '@mui/material';
-import Grid from '@mui/material/Unstable_Grid2'; // Grid version 2
-import { useMemo, useEffect, useCallback, useContext } from 'react';
+import { Paper, styled } from '@mui/material';
+import { useContext } from 'react';
 import ViewOptionsContext from '@/contexts/ViewOptionsContext';
-
-import useSupportedFilePicker from '@/hooks/useSupportedFilePicker';
-import { useHeldRepetitionTimer, useModelSelectionExport } from '@/hooks';
-import useSceneOBJFileDownloader from '@/hooks/useSceneOBJDownloader';
-import { mdiMenuLeftOutline, mdiMenuRightOutline } from '@mdi/js';
-import Icon from '@mdi/react';
 import clsx from 'clsx';
 import GuiPanelViewOptions from './GuiPanelViewOptions';
-import GuiPanelButton from './GuiPanelButton';
-import GuiPanelSection from './GuiPanelSection';
 import GuiPanelTextures from './GuiPanelTextures';
+import GuiPanelModels from './GuiPanelModels';
 
 const WIDTH = 222;
-
-// @TODO: consider either:
-// (1) breaking this panel into separate components,
-// (2) offload hook functionality since there's quite a lot of cruft
-// (3) abstract the components to eliminate cognitive overhead
 
 const TRANSITION_TIME = `0.32s`;
 
@@ -173,81 +139,7 @@ const StyledPaper = styled(Paper)(
 );
 
 export default function GuiPanel() {
-  // @TODO use a more standard error dialog vs using window.alert here
-  const openFileSelector = useSupportedFilePicker(globalThis.alert);
-
   const viewOptions = useContext(ViewOptionsContext);
-  const dispatch = useAppDispatch();
-  const onExportSelectionJson = useModelSelectionExport();
-  const onExportOBJFile = useSceneOBJFileDownloader();
-  const modelIndex = useAppSelector(selectModelIndex);
-  const modelCount = useAppSelector(selectModelCount);
-  const objectKey = useAppSelector(selectObjectKey);
-  const objectSelectionType = useAppSelector(selectObjectSelectionType);
-  const model = useAppSelector(selectModel);
-  const onSetObjectSelectionType = useCallback(
-    (_: React.MouseEvent<HTMLElement>, type: 'mesh' | 'polygon') => {
-      type && dispatch(setObjectType(type));
-    },
-    [objectSelectionType]
-  );
-
-  const exportSelectionButton = useMemo(() => {
-    {
-      let title = 'Export Model JSON';
-
-      if (objectKey) {
-        title = `Export ${
-          objectSelectionType === 'mesh' ? 'Mesh' : 'Polygon'
-        } JSON`;
-      }
-
-      return !model ? undefined : (
-        <Tooltip title='Export ModNao .json data. Will narrow data down to the current selection'>
-          <Button
-            fullWidth
-            onClick={onExportSelectionJson}
-            color='secondary'
-            size='small'
-            variant='outlined'
-          >
-            {title}
-          </Button>
-        </Tooltip>
-      );
-    }
-  }, [model, objectSelectionType, objectKey, onExportSelectionJson]);
-
-  const [onStartPrevModelNav, onStopPrevModelNav] = useHeldRepetitionTimer();
-  const [onStartNextModelNav, onStopNextModelNav] = useHeldRepetitionTimer();
-
-  useEffect(() => {
-    window.addEventListener('mouseup', onStopPrevModelNav);
-    window.addEventListener('mouseup', onStopNextModelNav);
-    return () => {
-      window.removeEventListener('mouseup', onStopPrevModelNav);
-      window.removeEventListener('mouseup', onStopNextModelNav);
-    };
-  }, []);
-
-  const onStartModelPrevClick = useCallback(() => {
-    onStartPrevModelNav(() => {
-      dispatch(navToPrevModel());
-    });
-  }, [modelIndex]);
-
-  const onStartModelNextClick = useCallback(() => {
-    onStartNextModelNav(() => {
-      dispatch(navToNextModel());
-    });
-  }, [modelIndex]);
-
-  let modelIndexAndCount = '--';
-
-  if (model) {
-    const sp = modelCount > 99 && modelIndex > 98 ? '' : ' ';
-    modelIndexAndCount = `${modelIndex + 1}${sp}/${sp}${modelCount}`;
-  }
 
   return (
     <StyledPaper
@@ -255,94 +147,7 @@ export default function GuiPanel() {
       className={clsx(viewOptions.guiPanelVisible && 'visible')}
     >
       <div className='content'>
-        <GuiPanelSection title='Models'>
-          <div className='selection'>
-            <Grid container className='property-table'>
-              <Grid xs={4}>
-                <Typography variant='body1' textAlign='right'>
-                  Models
-                </Typography>
-              </Grid>
-              <Grid xs={8}>
-                <IconButton
-                  className='model-nav-button'
-                  color='primary'
-                  aria-haspopup='true'
-                  onMouseDown={onStartModelPrevClick}
-                  onMouseUp={onStopPrevModelNav}
-                  disabled={!model || modelIndex === 0}
-                >
-                  <Icon path={mdiMenuLeftOutline} size={1} />
-                </IconButton>
-                <Typography variant='button' textAlign='right'>
-                  {modelIndexAndCount}
-                </Typography>
-                <IconButton
-                  className='model-nav-button'
-                  color='primary'
-                  aria-haspopup='true'
-                  onMouseDown={onStartModelNextClick}
-                  onMouseUp={onStopNextModelNav}
-                  disabled={!model || modelIndex === modelCount - 1}
-                >
-                  <Icon path={mdiMenuRightOutline} size={1} />
-                </IconButton>
-              </Grid>
-              <Grid xs={8}>
-                <Typography variant='body1' textAlign='right'>
-                  Object Key
-                </Typography>
-              </Grid>
-              <Grid xs={4}>
-                <Typography variant='button' textAlign='right'>
-                  {!objectKey ? '--' : objectKey}
-                </Typography>
-              </Grid>
-              <Grid xs={8}>
-                <Typography variant='body1' textAlign='right'>
-                  Object Type
-                </Typography>
-              </Grid>
-              <Grid xs={4}>
-                <ToggleButtonGroup
-                  orientation='vertical'
-                  color='secondary'
-                  value={objectSelectionType}
-                  size='small'
-                  exclusive
-                  onChange={onSetObjectSelectionType}
-                  aria-label='text alignment'
-                >
-                  <ToggleButton value='mesh'>mesh</ToggleButton>
-                  <ToggleButton value='polygon'>polygon</ToggleButton>
-                </ToggleButtonGroup>
-              </Grid>
-            </Grid>
-            <GuiPanelButton
-              tooltip='Select an MVC2 or CVS2 STG POL.BIN and/or TEX.BIN files'
-              onClick={openFileSelector}
-            >
-              Import Model/Texture
-            </GuiPanelButton>
-            {!model ? undefined : (
-              <GuiPanelButton
-                tooltip={
-                  <div>
-                    <p>
-                      Export an.obj file representing the selected in-scene
-                      model meshes.
-                    </p>
-                  </div>
-                }
-                onClick={onExportOBJFile}
-                color='secondary'
-              >
-                Export Model .OBJ
-              </GuiPanelButton>
-            )}
-            {exportSelectionButton}
-          </div>
-        </GuiPanelSection>
+        <GuiPanelModels />
         <GuiPanelTextures />
         <GuiPanelViewOptions />
       </div>
